fix UtkBase Uefi File parsing issues
UEFI Volume can have short padding at their end.
This was a problem when the padding is shorter then a FileHeader would need.
Funily enough, the debugger completely failed to break on any reasonable asserts i placed to debug this.
Maybe that is why i went a bit over-board with the exception-handling in this commit.