fix UtkBase UEFI Volume issue when no external header is present
It appears that when there is no External Header in a UEFI Volume, the VolumeHeaders offset for that external header, is 0.
In that case, none of the file offsets were calculated correctly.
This checks for bogus external header offsets.
Skipps bogus external header creations.
And properly defines the ContentOffset constant